wip: Implemented BlockchainWatcher (Ethereum) - Implement WithdrawalService

+++ src/main/java/com/zenz/neopay/api/route/withdrawal/response/WithdrawalResponse.java
@@ -3,12 +3,13 @@ package com.zenz.neopay.api.route.withdrawal.response;
 import com.zenz.neopay.enums.WithdrawalStatus;
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.UUID;
 
 @Data
 public class WithdrawalResponse {
     private UUID withdrawalId;
-    private long amount;
+    private BigInteger amount;
     private String currency;
     private String chain;
     private WithdrawalStatus status;
